requirements.txt
python-dotenv==1.1.1
selenium==4.34.0
undetected_chromedriver==3.5.5
webdriver_manager==4.0.1
